requirements.txt
# Core dependencies
uiautomator2>=3.0.0
Pillow>=10.0.0
customtkinter>=5.0.0

# Cloud AI APIs
google-generativeai>=0.8.0
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0

# Optional: Local model (legacy)
# ollama>=0.3.0
